app/requirements.txt
loguru==0.7.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.30
vk-botting==0.11.4